refactor(authorization): split authorization handlers in different classes
